Casting to ints produces incorrect results when the number is larger than 2^32.

+++ stepmania/src/arch/Sound/RageSoundDriver_Generic_Software.cpp
@@ -476,8 +476,8 @@ int64_t RageSoundDriver::ClampHardwareFrame( int64_t iHardwareFrame ) const
 		static RageTimer last(RageZeroTimer);
 		if( last.IsZero() || last.Ago() > 1.0f )
 		{
-			LOG->Trace( "RageSoundDriver: driver returned a lesser position (%i < %i)",
-				(int) iHardwareFrame, (int) m_iMaxHardwareFrame );
+			LOG->Trace( "RageSoundDriver: driver returned a lesser position (%lld < %lld)",
+				    iHardwareFrame, m_iMaxHardwareFrame );
 			last.Touch();
 		}
 		return m_iMaxHardwareFrame;
